+++ src/env.ts
@@ -1,1 +1,25 @@
-export const DATA_ENV: 'test' | 'prd' = 'prd';
+/**
+ * api domain for outer user
+ */
+export const PUBLIC_API = 'http://localhost:8085';
+
+/**
+ * api domain for internal user
+ */
+export const API = 'http://localhost:8085';
+
+/**
+ * login domain
+ */
+export const LOGIN = 'https://devpilot.zhongan.com';
+
+/**
+ * official site domain
+ */
+export const OFFICIAL_SITE = 'https://devpilot.zhongan.com';
+
+/**
+ * whether to check login
+ */
+export const AUTH_ON: boolean = true;
+export const TELEMETRY_ON: boolean = true;