+++ src/utils/domain.ts
@@ -1,8 +1,0 @@
-export default {
-  // api domain for outer user
-  PUBLIC_API: 'http://localhost:8085',
-  // api domain for internal user
-  API: 'http://localhost:8085',
-  // login domain（change to your own if needed）
-  LOGIN: 'https://devpilot.zhongan.com',
-};